backend/requirements.txt
functions-framework
flask
google-genai
google-cloud-vision
markdown